fix: photos テーブルが存在しない場合のエラー回避

+++ db/migrate/20260421035359_drop_photos_table.rb
@@ -1,5 +1,5 @@
 class DropPhotosTable < ActiveRecord::Migration[7.2]
   def change
-    drop_table :photos
+    drop_table :photos if table_exists?(:photos)
   end
 end
